refactor: inline proposal-summary helpers into safe-utils
Move getSigners, getSelector, summarizeProposalDoc and IProposalSummary
from the standalone proposal-status-utils module into safe-utils, next to
the existing signature helpers and ISafeTxDocument they operate on; move
their tests into safe-utils.test.ts. list-pending-proposals.ts now imports
them from safe-utils. Keeps all Safe-document helpers discoverable in one
place rather than a separate small module.
